dotbot: rename main.py -> controller_app.py
